Rename: components/table 에서 components/common 으로
