+++ components/common/RankBoard.vue
@@ -1,0 +1,7 @@
+<template>
+  <v-card class="rank-board d-flex px-8 py-3" max-width="254px"> </v-card>
+</template>
+<script>
+export default {}
+</script>
+<style lang="scss" scoped></style>
